Moved AeroCube-specific classes from Fiducial Marker Module to main ImP module
